cognitus_task/myproject/mydjangoapp/requirements.txt
anyio==3.6.1
certifi==2022.6.15
decorator==5.1.1
Django==4.0.6
django-crispy-forms==1.14.0
django-request==1.6.1
djangorestframework==3.13.1
executing==0.9.1
httptools==0.4.0
idna==3.3
ipdb==0.13.9
ipython==8.4.0
jedi==0.18.1
matplotlib-inline==0.1.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
ptyprocess==0.7.0
Pygments==2.12.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.11
uvicorn==0.18.2
websockets==10.3